2011-04-20  Maciej Stachowiak  <mjs@apple.com>
        Reviewed by Sam Weinig.

        Reproducible crash under WebPageProxy::didReceiveEvent
        https://bugs.webkit.org/show_bug.cgi?id=59036
        <rdar://problem/9252083>

        * UIProcess/API/mac/WKView.mm:
        (-[WKView _resendKeyDownEvent:]): Protect against self-destruction
        when re-dispatching the event.
        * UIProcess/WebProcessProxy.cpp:
        (WebKit::WebProcessProxy::createWebPage): Fixed a type mismatch
        I noticed while debugging.


Canonical link: https://commits.webkit.org/74173@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@84434 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/UIProcess/WebProcessProxy.cpp
@@ -169,7 +169,7 @@ PassRefPtr<WebPageProxy> WebProcessProxy::createWebPage(PageClient* pageClient, 
 {
     ASSERT(context->process() == this);
 
-    unsigned pageID = generatePageID();
+    uint64_t pageID = generatePageID();
     RefPtr<WebPageProxy> webPage = WebPageProxy::create(pageClient, this, pageGroup, pageID);
     m_pageMap.set(pageID, webPage.get());
     return webPage.release();
